[refactor] Update code structure, and component directory structure
